Fix crash in dirmngr when system certificates are loaded after fork().
GnuPG adds the system certificates for each gnutls connection when the
connection is configured. To read the system certificates, gnutls uses
Foundation methods. The use of Foundation methods however might lead to
unexpected behavior or worst a crash:
http://www.sealiesoftware.com/blog/archive/2017/6/5/Objective-C_and_fork_in_macOS_1013.html

To avoid this scenario, the certificates are read and cached before
fork() is called, and added to gnutls from the cache.

[#752]