gpgconf is always launched in verbose mode
A regression introduced in 2.2.33 (which is also included in the latest 2.2.34) causes gpgconf to always start in verbose mode, no matter whether or not the --verbose mode is passed. It's also no longer possible to disable verbosity with --no-verbose.
The following code is taken from the 2.3.x code base which doesn't include this bug.

A ticket in the GnuPG bug tracker was created: https://dev.gnupg.org/T5874

[#767]